code added for shmem_access()

+++ xv6_Unix/xv6-cuspring2019/sysproc.c
@@ -117,7 +117,7 @@ int sys_shmem_access(void)
       return 0; //returns nullptr
     }
 
-    return shmem_access(shmem_pgnum);
+    return (int)shmem_access(shmem_pgnum);
 }
